+++ lib/core/contents/index.js
@@ -6,7 +6,7 @@
 
 var Dao = require('../../dao/dao')
   , GeometryColumnsDao = require('../../dao/geometryColumns').GeometryColumnsDao
-  , SpatialReferenceSystemDao = require('../../dao/spatialReferenceSystem').SpatialReferenceSystemDao;
+  , SpatialReferenceSystemDao = require('../srs').SpatialReferenceSystemDao;
 
 var util = require('util');
