+++ lib/core/srs/index.js
@@ -4,7 +4,7 @@
  * @see module:dao/dao
  */
 
-var Dao = require('./dao');
+var Dao = require('../../dao/dao');
 var util = require('util');
 
 /**
