+++ lib/dao/geometryColumns.js
@@ -5,7 +5,7 @@
  */
 
 var Dao = require('./dao')
-  , SpatialReferenceSystemDao = require('./spatialReferenceSystem').SpatialReferenceSystemDao
+  , SpatialReferenceSystemDao = require('../core/srs').SpatialReferenceSystemDao
   , ContentsDao = require('../core/contents').ContentsDao
   , Contents = require('../core/contents').Contents;
