+++ lib/geoPackage.js
@@ -3,11 +3,11 @@
  * @module geoPackage
  */
 
-var SpatialReferenceSystemDao = require('./dao/spatialReferenceSystem').SpatialReferenceSystemDao
+var SpatialReferenceSystemDao = require('./core/srs').SpatialReferenceSystemDao
   , GeometryColumnsDao = require('./dao/geometryColumns').GeometryColumnsDao
   , FeatureDao = require('./features/user/featureDao')
   , FeatureTableReader = require('./features/user/featureTableReader')
-  , Contents = require('./core/contents').Contents;
+  , ContentsDao = require('./core/contents').ContentsDao;
 
 /**
  * GeoPackage database
@@ -40,7 +40,7 @@ GeoPackage.prototype.getSpatialReferenceSystemDao = function() {
 }
 
 GeoPackage.prototype.getContentsDao = function() {
-  return new Contents(this.db);
+  return new ContentsDao(this.db);
 }
 
 GeoPackage.prototype.createDao = function () {
