+++ lib/geoPackageManager.js
@@ -21,7 +21,9 @@ module.exports.open = function(filePath, callback) {
   var results = {};
   async.waterfall([function(callback) {
     results.db = new sqlite3.Database(filePath, function(err) {
+      console.log('err', err);
       if (err) {
+        console.log('cannot open ' + filePath);
         return callback(err);
       }
       callback(err, results);
@@ -34,6 +36,7 @@ module.exports.open = function(filePath, callback) {
       callback(err, results);
     });
   }], function(err, results) {
+    console.log('err', err);
     if (err) {
       if (results.db) {
         results.db.close();
