+++ lib/user/userDao.js
@@ -45,24 +45,10 @@ UserDao.prototype.getRow = function (results) {
   if (!this.table) return row;
 
   var columns = this.table.columnCount();
-  var columnTypes = new Array(columns);
-  var values = new Array(columns);
+  var columnTypes = new Array();
+  var values = new Array();
 
   return this.newRowWithColumnTypes(columnTypes, results);
-
-
-  // NSUInteger totalColumns = [self.columns count];
-  //
-  //   for (int i=0; i<totalColumns; i++){
-  //
-  //       if(types != nil){
-  //           [GPKGUtils addObject:[NSNumber numberWithInt:[self getType:i]] toArray:types];
-  //       }
-  //
-  //       NSObject * value = [self getValueWithIndex:i];
-  //       [GPKGUtils addObject:value toArray:values];
-  //   }
-
 };
 
 /**
