+++ lib/validate/geoPackageValidate.js
@@ -6,7 +6,7 @@
 
 var path = require('path')
   , async = require('async')
-  , SpatialReferenceSystem = require('../dao/spatialReferenceSystem').SpatialReferenceSystem
+  , SpatialReferenceSystem = require('../core/srs').SpatialReferenceSystem
   , Contents = require('../core/contents').Contents
   , GeoPackageConstants = require('../geoPackageConstants');
 
@@ -47,7 +47,9 @@ exports.hasMinimumTables = function(geoPackage, callback) {
       geoPackage.getSpatialReferenceSystemDao().isTableExists(callback);
     },
     function(callback) {
-      geoPackage.getContentsDao().isTableExists(callback);
+      var contentsDao = geoPackage.getContentsDao();
+      console.log('contents dao', contentsDao);
+      contentsDao.isTableExists(callback);
     }
   ], callback);
 }
